+++ flask_dance/__init__.py
@@ -1,3 +1,3 @@
 from .consumer import OAuth1ConsumerBlueprint, OAuth2ConsumerBlueprint
 
-__version__ = "5.0.0"
+__version__ = "5.1.0"